feat: DI refactor, structured errors, /v1/models, preloading, CORS, pre-commit
- Replace request.app.state access with FastAPI Depends (dependencies.py)
- Migrate API key auth from middleware to Depends with HTTPBearer (Swagger UI auth button)
- Add structured error responses with error_id, type, message (errors.py)
- Add model preloading via FUNASR_PRELOAD_MODELS config
- Add GET /v1/models endpoint (OpenAI-compatible model listing)
- Add configurable CORS middleware via FUNASR_ALLOW_ORIGINS
- Extract audio temp file handling into AudioFileDependency (yield-based cleanup)
- Add pre-commit config (ruff + pyright) and expand ruff rules
- All 55 tests passing, 97% coverage

+++ src/funasr_server/formatters.py
@@ -52,7 +52,7 @@ def _format_timestamp(seconds: float, *, vtt: bool = False) -> str:
         格式化后的时间戳字符串，如 "00:01:30,500" 或 "00:01:30.500"
     """
     # 先将秒转为总毫秒数（四舍五入），避免浮点精度问题
-    total_ms = int(round(seconds * 1000))
+    total_ms = round(seconds * 1000)
 
     # 依次拆分出 时、分、秒、毫秒（使用 divmod 整除取余）
     hours, remainder = divmod(total_ms, 3_600_000)  # 1 小时 = 3,600,000 毫秒
